reduced number of single component files
