+++ src/modules/comment/CommentsList/CommentsList.tsx
@@ -1,6 +1,23 @@
 import { component$ } from "@builder.io/qwik";
 import type { Comment } from "@prisma/client";
-import { CommentsListItem } from "./CommentsListItem/CommentsListItem";
+import { paths } from "~/utils/paths";
+import { CommentActions } from "../CommentActions/CommentActions";
+
+type CommentsListItemProps = {
+  comment: Comment;
+};
+
+export const CommentsListItem = component$<CommentsListItemProps>((props) => {
+  return (
+    <div>
+      <pre>{JSON.stringify(props.comment, null, 2)}</pre>
+      <a class="link" href={paths.comment(props.comment.id)}>
+        Show comments
+      </a>
+      <CommentActions comment={props.comment} />
+    </div>
+  );
+});
 
 type Props = {
   comments: Comment[];
