+++ src/routes/board/comment/[commentId]/index.tsx
@@ -1,7 +1,11 @@
 import { component$ } from "@builder.io/qwik";
 import { routeLoader$, type DocumentHead } from "@builder.io/qwik-city";
+import type { Comment } from "@prisma/client";
+import { CommentActions } from "~/modules/comment/CommentActions/CommentActions";
+import { CommentsList } from "~/modules/comment/CommentsList/CommentsList";
+import { CreateCommentForm } from "~/modules/comment/CreateCommentForm/CreateCommentForm";
 import { trpc } from "~/server/trpc/api";
-import { CommentCard } from "./CommentCard/CommentCard";
+import { paths } from "~/utils/paths";
 
 export const useComment = routeLoader$((event) =>
   trpc.comment.get.loader(event, { id: event.params.commentId })
@@ -14,6 +18,38 @@ export const useComments = routeLoader$((event) =>
     take: 10,
   })
 );
+
+type CommentCardProps = {
+  comment: Comment;
+};
+
+export const CommentCard = component$<CommentCardProps>((props) => {
+  const comments = useComments();
+
+  const backPath = props.comment.parentId
+    ? paths.comment(props.comment.parentId)
+    : paths.post(props.comment.postId);
+
+  return (
+    <div>
+      <a class="link" href={backPath}>
+        Back
+      </a>
+      <pre>{JSON.stringify(props.comment, null, 2)}</pre>
+      <CommentActions comment={props.comment} />
+      <CreateCommentForm
+        parentId={props.comment.id}
+        postId={props.comment.postId}
+      />
+      {comments.value.status === "success" ? (
+        <CommentsList
+          comments={comments.value.result.comments}
+          count={comments.value.result.count}
+        />
+      ) : null}
+    </div>
+  );
+});
 
 export default component$(() => {
   const comment = useComment();
