--- src/routes/board/post/[postId]/PostCard/PostCard.tsx
@@ -1,32 +1,0 @@
-import { component$ } from "@builder.io/qwik";
-import type { Post } from "@prisma/client";
-import { CommentsList } from "~/modules/comment/CommentsList/CommentsList";
-import { CreateCommentForm } from "~/modules/comment/CreateCommentForm/CreateCommentForm";
-import { PostActions } from "~/modules/post/PostActions/PostActions";
-import { paths } from "~/utils/paths";
-import { useComments } from "..";
-
-type Props = {
-  post: Post;
-};
-
-export const PostCard = component$<Props>((props) => {
-  const comments = useComments();
-
-  return (
-    <div>
-      <a class="link" href={paths.board}>
-        Back
-      </a>
-      <pre>{JSON.stringify(props.post, null, 2)}</pre>
-      <PostActions post={props.post} />
-      <CreateCommentForm parentId={null} postId={props.post.id} />
-      {comments.value.status === "success" ? (
-        <CommentsList
-          comments={comments.value.result.comments}
-          count={comments.value.result.count}
-        />
-      ) : null}
-    </div>
-  );
-});
